+++ fake-cucumber/javascript/src/ExpressionStepDefinition.ts
@@ -1,7 +1,6 @@
 import ISupportCodeExecutor from './ISupportCodeExecutor'
 import SupportCodeExecutor from './SupportCodeExecutor'
 import {
-  Argument,
   CucumberExpression,
   Expression,
   RegularExpression,
@@ -21,7 +20,7 @@ export default class ExpressionStepDefinition implements IStepDefinition {
   public match(
     pickleStep: messages.Pickle.IPickleStep
   ): ISupportCodeExecutor | null {
-    const expressionArgs = this.getArguments(pickleStep.text)
+    const expressionArgs = this.expression.match(pickleStep.text)
     return expressionArgs === null
       ? null
       : new SupportCodeExecutor(
@@ -31,10 +30,6 @@ export default class ExpressionStepDefinition implements IStepDefinition {
           pickleStep.argument && pickleStep.argument.docString,
           pickleStep.argument && pickleStep.argument.dataTable
         )
-  }
-
-  public getArguments(text: string): ReadonlyArray<Argument<any>> {
-    return this.expression.match(text)
   }
 
   public toMessage(): messages.IEnvelope {