+++ fake-cucumber/javascript/src/cli.ts
@@ -14,11 +14,6 @@ program.version(packageJson.version)
 program.option('-r, --require <path>', 'override require path')
 program.option('--predictable-ids', 'Use predictable ids', false)
 program.option(
-  '--globals',
-  'Assign Given/When/Then/After/Before to global scope',
-  false
-)
-program.option(
   '-f, --format <format>',
   'output format: ndjson|protobuf',
   'protobuf'
@@ -26,16 +21,12 @@ program.option(
 
 async function main() {
   program.parse(process.argv)
-  const { globals, predictableIds, format } = program
+  const { predictableIds, format } = program
 
   const paths = program.args
   const requirePaths = program.require ? program.require.split(':') : paths
 
-  const supportCode = await loadSupportCode(
-    predictableIds,
-    requirePaths,
-    globals
-  )
+  const supportCode = await loadSupportCode(predictableIds, requirePaths)
 
   const gherkinOptions: IGherkinOptions = {
     defaultDialect: 'en',