+++ fake-cucumber/javascript/src/loadSupportCode.ts
@@ -14,16 +14,9 @@ import IClock from './IClock'
 
 export default async function loadSupportCode(
   predictableIds: boolean,
-  requirePaths: ReadonlyArray<string>,
-  globals: boolean
+  requirePaths: ReadonlyArray<string>
 ): Promise<SupportCode> {
   const supportCode = makeSupportCode(predictableIds)
-  if (globals) {
-    for (const key of Object.keys(dsl)) {
-      // @ts-ignore
-      global[key] = supportCode[key]
-    }
-  }
 
   dsl.setSupportCode(supportCode)
   const supportCodePaths = await findSupportCodePaths(requirePaths)